Adjust dockerfile to maintain xbps repodata
Adds services for xq-api and xbps-update. xbps-update just runs
sync-xbps with a handful of architectures so that they can all be
queried _maybe_.

There's an initial run of sync-xbps in the dockerfile to at least warm
it up because the first time it runs sync, it'll probably be a while
after starting the container (didn't want to run it on startup to avoid
restart loops being jerks).