Try using bleve for search indexing
This is _usually_ a lot faster than the dumb lowercase search that is
otherwise done for queries. It produces different results, but this
seems like a reasonable tradeoff to me.

Pagination will work with this, but is currently turned off because
I need to figure out how the frontend will work with it.

The package mapping is a giant hack because I need to figure out more
about how to properly set up mappings in bleve. The english analyzer
and such is really not ideal for a lot of these fields.