Add experimental frontend in Elm
Adds an experiment frontend written in Elm. This is not currently
vendored using npm or anything -- that'll change later once I work out
automation here, since elm kind of functions on its own package manager
anyway.

This is intended to replace the work-in-progress vue.js frontend that's
currently available at https://xbps.spiff.io. This is because even if
Vue is less complicated than other options, it's still complicated.
Comparatively, having a friendly language with a friendly compiler and
friendly runtime goes a long way towards making this easier to work on.

This is subject to change or be removed.