Clean up program entrypoint
This adds signal handling for shutdown, which is only really necessary
for Unix domain sockets (to remove them on close). These only take
effect after the initial repodata loading -- prior to that, a sigint or
sigterm will kill the process on the spot. This is to grant an early
out while the process is booting and has nothing to possibly clean up.

In addition, this adds environment variables to configure a subset of
available flags. These are only used in Docker currently.

An additional CLI flag, -net <type>, is added to support Unix domain
sockets as a listener. You can realistically pass anything supported by
net.Listen, but not all of them are likely to be useful for an HTTP
server (not that I'll stop you).