Grab repository name from repodata path
This fills a hole in some data missing from the older package search
since it was able to use file lists instead of repodata.

Currently, this will default to looking for _current_ in the directory
name of the repodata and, if present, everything after it will be
presented as the repo after replacing all underscores with slashes. So,
*_current_multilib_nonfree will become multilib/nonfree. If the
_current_ separator is not found, it selects the last
underscore-separated word from the directory name and uses that (this
matches "current", typically). Finally, if no name can be found in the
path, it defaults to "current".

In addition, there's some additional log output at verbosity 1 and
2. At level 1, when repodata is done being scanned it will print the
number of new packages that were parsed and the first path being
scanned. At level 2, it will also print an info message for every
directory that is scanned recursively.