Use dir names as repo name for scanned repodata
When scanning repository trees for repodata, use the suffix of the
repodata (after the search root) as the repository name. This breaks
older behavior based on local repodata pulled by xbps and is instead
focused on being used to search void-packages hostdirs.

The old behavior is retained when specifying individual repodata files
(because there is no search root) on the command line or when the
search root is "/var/db/xbps". For any other directory search, the new
behavior is used.

Fixes #5.