Add -reload-every flag (XQAPI_RELOAD_EVERY)
Add a -reload-every flag to reload the xq-api's repodata on a fixed
interval. By default, this is disabled. The value of the flag can be
set by setting the XQAPI_RELOAD_EVERY environment variable.

The logic for reloading on HUP has been moved down to the
reloadOnSignal function, to keep the triggered reload logic out of the
main function. Figured that now that there's two cases of it, it would
be best to get that out of the way now.